requirements.txt
streamlit==1.37.1
pyrebase4==4.7.1
firebase-admin==6.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==2.18.2
google-auth==2.33.0
google-generativeai==0.7.2
reportlab==4.2.2
pandas==2.2.2
numpy==1.26.4
plotly==5.23.0
altair==5.4.1
python-dateutil==2.9.0.post0